found a comprimise between betty and functionality betty better agree on this comprimise

+++ bit_manipulation/1-print_binary.c
@@ -8,8 +8,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long int max = (unsigned long int) 1 << (
-		sizeof(unsigned long int) * 8 - 1);
+	unsigned long int max = 1UL << (sizeof(unsigned long int) * 8 - 1);
 	int started = 0;
 
 	if (n == 0)
